requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
pymongo
google-generativeai
textblob